[WalletProvider] move Portal and Dialog to containers
