+++ packages/react-wallet/src/containers/Dialog/index.tsx
